pyproject.toml
[project]
name = "Async-vs-Thread-Scraping"
version = "0.1.0"
description = "Benchmarking synchronous and asynchronous HTTP requests."
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "aiohttp>=3.11.18",
    "fastapi>=0.115.12",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
